First commit for Mythic Depths
Really this is just the cookiecutter template for a pygame project (from
Microsoft) plus an example from the pygame website and a basic RPG class
definition from ChatGPT. But ya gotta start somewhere, right?

The goal here is to make a dungeon crawler with stairs at the end of
each "room" of the dungeon, where you climb deeper into the depths.
Procedural generation is something I want to learn more about as I make
this project.

README.md contains goals for the project, which will change greatly over
time. I'm sure the stuff in here will be completely revamped after a
commit, but wanted to get something on the books.